feat: criando method's para cidade/estado

+++ src/main/java/com/algaworks/algafood/infrastructure/repository/CozinhaRepositoryImpl.java
@@ -4,6 +4,7 @@ import com.algaworks.algafood.domain.model.Cozinha;
 import com.algaworks.algafood.domain.repository.CozinhaRepository;
 import jakarta.persistence.EntityManager;
 import jakarta.persistence.PersistenceContext;
+import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Component;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -16,13 +17,13 @@ public class CozinhaRepositoryImpl implements CozinhaRepository {
     private EntityManager entityManager;
 
     @Override
-    public List<Cozinha> todas() {
+    public List<Cozinha> listar() {
         return entityManager.createQuery("from Cozinha", Cozinha.class).getResultList();
 
     }
 
     @Override
-    public Cozinha porId(Long id) {
+    public Cozinha buscar(Long id) {
         return entityManager.find(Cozinha.class, id);
     }
 
@@ -34,8 +35,10 @@ public class CozinhaRepositoryImpl implements CozinhaRepository {
 
     @Override
     @Transactional
-    public void remover(Cozinha cozinha) {
-        cozinha = porId(cozinha.getId());
+    public void remover(Long id) {
+        Cozinha cozinha = buscar(id);
+        if (cozinha == null) throw new EmptyResultDataAccessException(1);
+        
         entityManager.remove(cozinha);
     }
 }
